Move verifier to new crate, adjust location of dev/prod, sw/hw feature
