+++ Cargo.toml
@@ -14,6 +14,7 @@ members = [
     "attest/net",
     "attest/trusted",
     "attest/untrusted",
+    "attest/verifier",
     "common",
     "connection",
     "connection/test-utils",
