+++ attest/core/Cargo.toml
@@ -17,14 +17,8 @@ std = [
     "mc-common/std",
     "digest/std",
     "displaydoc/std",
-    "mbedtls/std",
     "sha2/std",
 ]
-
-# This means "use the fake SGX stuff where required"
-sgx-sim = []
-# This means "use the dev endpoint of Intel's Attestation Service"
-ias-dev = []
 
 [dependencies]
 mc-common = { path = "../../common", default-features = false }
@@ -33,8 +27,6 @@ mc-crypto-rand = { path = "../../crypto/rand" }
 mc-util-encodings = { path = "../../util/encodings" }
 mc-sgx-css = { path = "../../sgx/css" }
 mc-sgx-types = { path = "../../sgx/types" }
-
-mbedtls = { version = "0.8.1", default-features = false, features = ["no_std_deps"] }
 
 binascii = "0.1.2"
 bitflags = "1.2"
@@ -51,6 +43,7 @@ prost = { version = "0.9", default-features = false }
 
 [dev-dependencies]
 mc-util-serial = { path = "../../util/serial" }
+
 bincode = "1.3"
 hex = "0.4"
 pem = "1.0"
@@ -58,15 +51,7 @@ rand = "0.8"
 rand_hc = "0.3"
 
 [build-dependencies]
-mc-sgx-build =  { path = "../../sgx/build" }
+mc-util-build-script = { path = "../../util/build/script" }
+mc-util-build-sgx = { path = "../../util/build/sgx" }
 
-cargo-emit = "0.2"
-chrono = "0.4"
-hex = "0.4"
-lazy_static = "1.4"
-rand = "0.8"
-rand_hc = "0.3"
-
-# We use mbedtls to generate certificates for simulation mode
-mbedtls = "0.8.1"
-mbedtls-sys-auto = "2.26.1"
+cargo-emit = "0.2.1"
