+++ attest/core/src/ias.rs
@@ -3,8 +3,4 @@
 //! Intel Attestation Service Support
 
 pub mod json;
-pub mod verifier;
 pub mod verify;
-
-#[cfg(feature = "sgx-sim")]
-pub mod sim;
