+++ attest/core/src/lib.rs
@@ -27,11 +27,8 @@ pub use crate::{
         QuoteVerifyError, ReportBodyVerifyError, ReportDetailsError, RevocationCause, SgxError,
         SgxResult, SignatureError, TargetInfoError, VerifyError,
     },
-    ias::{
-        verifier::{Error as VerifierError, MrEnclaveVerifier, MrSignerVerifier, Verifier},
-        verify::{
-            EpidPseudonym, VerificationReport, VerificationReportData, VerificationSignature,
-        },
+    ias::verify::{
+        EpidPseudonym, VerificationReport, VerificationReportData, VerificationSignature,
     },
     nonce::{IasNonce, Nonce, QuoteNonce},
     quote::{Quote, QuoteSignType},
@@ -59,33 +56,5 @@ pub use crate::{
     },
 };
 
-#[cfg(feature = "sgx-sim")]
-pub use crate::ias::sim::{
-    IAS_SIM_MODULUS, IAS_SIM_ROOT_ANCHORS, IAS_SIM_SIGNING_CHAIN, IAS_SIM_SIGNING_KEY,
-};
-
 /// The IAS version we support
 pub const IAS_VERSION: f64 = 4.0;
-
-cfg_if! {
-    if #[cfg(feature = "sgx-sim")] {
-        /// Whether or not enclaves should be run and validated in debug mode
-        pub const DEBUG_ENCLAVE: bool = true;
-        /// An array of zero-terminated signing certificate PEM files used as root anchors.
-        pub const IAS_SIGNING_ROOT_CERT_PEMS: &[&str] = &[crate::IAS_SIM_ROOT_ANCHORS];
-    } else if #[cfg(feature = "ias-dev")] {
-        /// Whether or not enclaves should be run and validated in debug mode
-        pub const DEBUG_ENCLAVE: bool = true;
-        /// An array of zero-terminated signing certificate PEM files used as root anchors.
-        pub const IAS_SIGNING_ROOT_CERT_PEMS: &[&str] = &[concat!(include_str!(
-            "../data/Dev_AttestationReportSigningCACert.pem"
-        ), "\0")];
-    } else {
-        /// Debug enclaves in prod mode are not supported.
-        pub const DEBUG_ENCLAVE: bool = false;
-        /// An array of zero-terminated signing certificate PEM files used as root anchors.
-        pub const IAS_SIGNING_ROOT_CERT_PEMS: &[&str] = &[concat!(include_str!(
-            "../data/AttestationReportSigningCACert.pem"
-        ), "\0")];
-    }
-}
