+++ consensus/enclave/trusted/Cargo.toml
@@ -11,14 +11,15 @@ crate-type = ["staticlib"]
 
 [features]
 sgx-sim = [
-    "mc-attest-core/sgx-sim"
+    "mc-attest-verifier/sgx-sim"
 ]
 ias-dev = [
-    "mc-attest-core/ias-dev"
+    "mc-attest-verifier/ias-dev"
 ]
 
 [dependencies]
 mc-attest-core = { path = "../../../attest/core", default-features = false }
+mc-attest-verifier = { path = "../../../attest/verifier", default-features = false }
 mc-attest-trusted = { path = "../../../attest/trusted" }
 mc-common = { path = "../../../common", default-features = false }
 mc-consensus-enclave-api = { path = "../api", default-features = false }
